backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
openai==1.6.1
anthropic==0.18.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
